Add the manual, downloaded from MOCAGH.
http://mocagh.org/loadpage.php?getgame=hexx

The manual was needed to play the original game (the icons in
the corners of the pages are an anti-piracy measure), but the
Abandonia version of the game does not have the anti-piracy
gate at the beginning. The manual is still useful for learning
the interface, learning when to level up, and learning the
spells' and magic items' effects.